move parser back to main
